renamed dir/ and fixed scripts
